+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # RELEASE NOTES
 ## 1.1
-
+- Fixed [#108](https://github.com/spetzel2020/quick-encounters/issues/108): Removed Hide button in QE because of multi-page Journals and always add Show button (for when you have Show QE Automatically set false)
 ## 1.0
 - Issue #71: Instant Encounters: Drag a Quick Encounter directly from the Journals directory to the Scene to run it!
 ### 1.0.8